Restructure, cefactor code, introduce services, and update readme
