+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "jest"
+    "test": "jest",
+    "start": "ts-node src/index.ts"
   },
   "author": "",
   "license": "ISC",
@@ -19,7 +20,8 @@
   "dependencies": {
     "axios": "^0.21.1",
     "dotenv": "^8.2.0",
-    "yarn": "^1.22.22"
+    "yarn": "^1.22.22",
+    "ts-node": "10.9.2"
   },
   "jest": {
     "testEnvironment": "node"
