+++ src/services/logger.ts
@@ -1,0 +1,9 @@
+export class Logger {
+    log(message: string): void {
+        console.log(`[INFO] ${message}`);
+    }
+
+    error(message: string, error: any): void {
+        console.error(`[ERROR] ${message}`, error);
+    }
+}
